Refactoring: create pdfService for server.
