+++ server/pdfService.js
@@ -1,0 +1,87 @@
+/**
+ * Service - PDF related.
+ */
+const path = require('path');
+
+/**
+ * Save a PDF file, and return the saved path.
+ */
+module.export = function savePDF(fileName, content) {
+    return new Promise((resolve, reject) => {
+
+        const dataPath = path.resolve(__dirname, 'server-data');
+        if (!fs.existsSync(dataPath)) {
+                fs.mkdirSync(dataPath);
+        }
+        const pdfPath = path.resolve(dataPath, fileName);
+        fs.writeFileSync(pdfPath, content);
+
+        resolve(pdfPath);
+    });
+}
+
+// Analize pdf with pdfreader.jar.
+module.export = function analyzePDF(pdfPath) {
+
+    return new Promise((resolve, reject) => {
+
+        // Check java command exits.
+        execCommand('java -version')
+            .then(resolve)
+            .catch(() => {
+                reject('java command not found.');
+            });
+
+    }).then(() => {
+
+        // Prepare pdfreader.jar
+
+        const exists = fs.existsSync(path.resolve(__dirname, 'pdfreader.jar'));
+        if (exists) {
+            return;
+        }
+
+        return new Promise((resolve, reject) => {
+
+            const reqConfig = {
+                method   : 'GET',
+                url      : 'https://cl.naist.jp/~shindo/pdfreader.jar',
+                encoding : null
+            };
+
+            request(reqConfig, function(err, response, buf) {
+
+                if (err) {
+                    reject(err);
+                }
+
+                fs.writeFileSync(path.resolve(__dirname, 'pdfreader.jar'), buf);
+
+                resolve();
+            });
+        });
+
+    }).then(() => {
+
+        const jarPath = path.resolve(__dirname, 'pdfreader.jar');
+        const cmd = `java -classpath ${jarPath} TextDrawImageExtractor ${pdfPath}`;
+        console.log('cmd:', cmd);
+        return execCommand(cmd);
+
+    }).then(({ stdout, stderr }) => {
+
+        return stdout;
+    });
+}
+
+// Execute an external command.
+function execCommand(command) {
+    return new Promise((resolve, reject) => {
+        exec(command, { maxBuffer : 1024 * 1024 * 50 }, (err, stdout, stderr) => {
+            if (err) {
+                reject({ err, stdout, stderr });
+            }
+            resolve({ stdout, stderr });
+        });
+    });
+}
